+++ util/ajax.js
@@ -94,8 +94,9 @@ function ajax(opt){
                 error: xhr.statusText||'no_status', message: xhr.responseText,
             };
         }
-        err.hola_info = {url: url, method: method, status: xhr.status,
+        err.xhr_info = {url: url, method: method, status: xhr.status,
             data: get_res_data(xhr), response_text: xhr.responseText};
+        err.hola_info = err.xhr_info; // legacy
         err.x_error = xhr.getResponseHeader('X-Luminati-Error') ||
             xhr.getResponseHeader('X-Hola-Error');
         throw err;